feat: evolutionフィーチャーのコンポーネント分割を実装
- PokemonCompareView, EvolutionActionButtons, EvolutionResultContentをコンポーネント化
- packages/app/lib/ui/features/evolution/components/配下に移動
- 設定クラス(Config)を使用したパラメータ管理を導入
- evolution_confirmation_page.dartとevolution_result_page.dartから大きなWidgetクラスを削除し、新しいコンポーネントを使用
